+++ src/interruptible_stream_ext.rs
@@ -65,6 +65,10 @@ pub trait InterruptibleStreamExt {
     ///     interruptible_stream.next().await
     /// );
     /// assert_eq!(
+    ///     Some(PollOutcome::NoInterrupt(1)),
+    ///     interruptible_stream.next().await
+    /// );
+    /// assert_eq!(
     ///     Some(PollOutcome::Interrupted(None)),
     ///     interruptible_stream.next().await
     /// );
@@ -161,9 +165,12 @@ where
 #[cfg(test)]
 mod tests {
     use futures::{stream, StreamExt};
-    use tokio::sync::{
-        mpsc::{self, error::SendError},
-        oneshot,
+    use tokio::{
+        sync::{
+            mpsc::{self, error::SendError},
+            oneshot,
+        },
+        task::yield_now,
     };
 
     use super::InterruptibleStreamExt;
@@ -322,6 +329,10 @@ mod tests {
             interruptible_stream.next().await
         );
         assert_eq!(
+            Some(PollOutcome::NoInterrupt(1u32)),
+            interruptible_stream.next().await
+        );
+        assert_eq!(
             Some(PollOutcome::Interrupted(None)),
             interruptible_stream.next().await
         );
@@ -329,13 +340,35 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_before() {
+    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_before_0() {
         let (interrupt_tx, interrupt_rx) = mpsc::channel::<InterruptSignal>(16);
 
         let mut interruptible_stream = stream::unfold(0u32, move |n| async move {
             if n < 3 { Some((n, n + 1)) } else { None }
         })
-        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 2).into());
+        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 0).into());
+
+        interrupt_tx
+            .send(InterruptSignal)
+            .await
+            .expect("Expected to send `InterruptSignal`.");
+
+        // interruption polled here, so `None` is returned
+        assert_eq!(
+            Some(PollOutcome::Interrupted(None)),
+            interruptible_stream.next().await
+        );
+        assert_eq!(None, interruptible_stream.next().await);
+    }
+
+    #[tokio::test]
+    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_before_1() {
+        let (interrupt_tx, interrupt_rx) = mpsc::channel::<InterruptSignal>(16);
+
+        let mut interruptible_stream = stream::unfold(0u32, move |n| async move {
+            if n < 3 { Some((n, n + 1)) } else { None }
+        })
+        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 1).into());
 
         interrupt_tx
             .send(InterruptSignal)
@@ -356,7 +389,76 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_between() {
+    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_before_2() {
+        let (interrupt_tx, interrupt_rx) = mpsc::channel::<InterruptSignal>(16);
+
+        let mut interruptible_stream = stream::unfold(0u32, move |n| async move {
+            if n < 3 { Some((n, n + 1)) } else { None }
+        })
+        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 2).into());
+
+        interrupt_tx
+            .send(InterruptSignal)
+            .await
+            .expect("Expected to send `InterruptSignal`.");
+
+        // interruption polled here
+        assert_eq!(
+            Some(PollOutcome::NoInterrupt(0u32)),
+            interruptible_stream.next().await
+        );
+        // interruption also polled here
+        assert_eq!(
+            Some(PollOutcome::NoInterrupt(1u32)),
+            interruptible_stream.next().await
+        );
+        // interruption also polled here, so `None` is returned
+        assert_eq!(
+            Some(PollOutcome::Interrupted(None)),
+            interruptible_stream.next().await
+        );
+        assert_eq!(None, interruptible_stream.next().await);
+    }
+
+    #[tokio::test]
+    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_before_3() {
+        let (interrupt_tx, interrupt_rx) = mpsc::channel::<InterruptSignal>(16);
+
+        let mut interruptible_stream = stream::unfold(0u32, move |n| async move {
+            if n < 3 { Some((n, n + 1)) } else { None }
+        })
+        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 3).into());
+
+        interrupt_tx
+            .send(InterruptSignal)
+            .await
+            .expect("Expected to send `InterruptSignal`.");
+
+        // interruption polled here
+        assert_eq!(
+            Some(PollOutcome::NoInterrupt(0u32)),
+            interruptible_stream.next().await
+        );
+        // interruption also polled here
+        assert_eq!(
+            Some(PollOutcome::NoInterrupt(1u32)),
+            interruptible_stream.next().await
+        );
+        // interruption also polled here
+        assert_eq!(
+            Some(PollOutcome::NoInterrupt(2u32)),
+            interruptible_stream.next().await
+        );
+        // interruption also polled here, so `None` is returned
+        assert_eq!(
+            Some(PollOutcome::Interrupted(None)),
+            interruptible_stream.next().await
+        );
+        assert_eq!(None, interruptible_stream.next().await);
+    }
+
+    #[tokio::test]
+    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_between_0() {
         let (interrupt_tx, interrupt_rx) = mpsc::channel::<InterruptSignal>(16);
         let (interrupt_ready_tx, mut interrupt_ready_rx) = mpsc::channel::<()>(2);
         let (interrupted_tx, interrupted_rx) = oneshot::channel::<()>();
@@ -364,9 +466,6 @@ mod tests {
         let mut interruptible_stream = stream::unfold(
             (0u32, Some((interrupt_ready_tx, interrupted_rx))),
             move |(n, channel_tx_rx)| async move {
-                if n == 0 {
-                    return Some((0, (1, channel_tx_rx)));
-                }
                 if let Some((interrupt_ready_tx, interrupted_rx)) = channel_tx_rx {
                     interrupt_ready_tx
                         .send(())
@@ -383,7 +482,72 @@ mod tests {
                 }
             },
         )
-        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 2).into());
+        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 0).into());
+
+        let interrupt_task = async {
+            interrupt_ready_rx
+                .recv()
+                .await
+                .expect("Expected `interrupt_ready_rx`. to receive message.");
+            interrupt_tx
+                .send(InterruptSignal)
+                .await
+                .expect("Expected to send `InterruptSignal`.");
+            interrupted_tx
+                .send(())
+                .expect("Expected to notify future to return value.");
+        };
+
+        let ((poll_outcome_first, poll_outcome_second), ()) = tokio::join!(
+            async {
+                (
+                    interruptible_stream.next().await,
+                    interruptible_stream.next().await,
+                )
+            },
+            interrupt_task
+        );
+
+        // First item is not interrupted.
+        assert_eq!(
+            Some(PollOutcome::Interrupted(Some(0u32))),
+            poll_outcome_first
+        );
+        // Second item is `None`, as `PollNextN`'s value is used up by the
+        // interruption.
+        assert_eq!(None, poll_outcome_second);
+    }
+
+    #[tokio::test]
+    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_between_1() {
+        let (interrupt_tx, interrupt_rx) = mpsc::channel::<InterruptSignal>(16);
+        let (interrupt_ready_tx, mut interrupt_ready_rx) = mpsc::channel::<()>(2);
+        let (interrupted_tx, interrupted_rx) = oneshot::channel::<()>();
+
+        let mut interruptible_stream = stream::unfold(
+            (0u32, Some((interrupt_ready_tx, interrupted_rx))),
+            move |(n, channel_tx_rx)| async move {
+                if n == 0 {
+                    return Some((n, (n + 1, channel_tx_rx)));
+                }
+                if let Some((interrupt_ready_tx, interrupted_rx)) = channel_tx_rx {
+                    interrupt_ready_tx
+                        .send(())
+                        .await
+                        .expect("Expected to send to interrupt ready channel.");
+                    let () = interrupted_rx
+                        .await
+                        .expect("Expected to be notified to return value.");
+                    yield_now().await;
+                }
+                if n < 3 {
+                    Some((n, (n + 1, None)))
+                } else {
+                    None
+                }
+            },
+        )
+        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 1).into());
 
         let interrupt_task = async {
             interrupt_ready_rx
@@ -411,13 +575,81 @@ mod tests {
 
         // First item is not interrupted.
         assert_eq!(Some(PollOutcome::NoInterrupt(0u32)), poll_outcome_first);
+        // Second item is `Interrupted`, as `PollNextN`'s second value is used up by the
+        // interruption.
+        assert_eq!(
+            Some(PollOutcome::Interrupted(Some(1u32))),
+            poll_outcome_second
+        );
+        assert_eq!(None, interruptible_stream.next().await);
+    }
+
+    #[tokio::test]
+    async fn interrupt_with_poll_next_n_returns_n_items_variant_interrupt_between_2() {
+        let (interrupt_tx, interrupt_rx) = mpsc::channel::<InterruptSignal>(16);
+        let (interrupt_ready_tx, mut interrupt_ready_rx) = mpsc::channel::<()>(2);
+        let (interrupted_tx, interrupted_rx) = oneshot::channel::<()>();
+
+        let mut interruptible_stream = stream::unfold(
+            (0u32, Some((interrupt_ready_tx, interrupted_rx))),
+            move |(n, channel_tx_rx)| async move {
+                if n <= 1 {
+                    return Some((n, (n + 1, channel_tx_rx)));
+                }
+                if let Some((interrupt_ready_tx, interrupted_rx)) = channel_tx_rx {
+                    interrupt_ready_tx
+                        .send(())
+                        .await
+                        .expect("Expected to send to interrupt ready channel.");
+                    let () = interrupted_rx
+                        .await
+                        .expect("Expected to be notified to return value.");
+                    yield_now().await;
+                    yield_now().await;
+                }
+                if n < 3 {
+                    Some((n, (n + 1, None)))
+                } else {
+                    None
+                }
+            },
+        )
+        .interruptible_with(Interruptibility::poll_next_n(interrupt_rx.into(), 2).into());
+
+        let interrupt_task = async {
+            interrupt_ready_rx
+                .recv()
+                .await
+                .expect("Expected `interrupt_ready_rx`. to receive message.");
+            interrupt_tx
+                .send(InterruptSignal)
+                .await
+                .expect("Expected to send `InterruptSignal`.");
+            interrupted_tx
+                .send(())
+                .expect("Expected to notify future to return value.");
+        };
+
+        let ((poll_outcome_first, poll_outcome_second, poll_outcome_third), ()) = tokio::join!(
+            async {
+                (
+                    interruptible_stream.next().await,
+                    interruptible_stream.next().await,
+                    interruptible_stream.next().await,
+                )
+            },
+            interrupt_task
+        );
+
+        // First item is not interrupted.
+        assert_eq!(Some(PollOutcome::NoInterrupt(0u32)), poll_outcome_first);
         // Second item is not interrupted, uses 1 of `PollNextN`.
         assert_eq!(Some(PollOutcome::NoInterrupt(1u32)), poll_outcome_second);
         // Third item is `None`, as `PollNextN`'s second value is used up by the
         // interruption.
         assert_eq!(
-            Some(PollOutcome::Interrupted(None)),
-            interruptible_stream.next().await
+            Some(PollOutcome::Interrupted(Some(2u32))),
+            poll_outcome_third
         );
         assert_eq!(None, interruptible_stream.next().await);
     }