+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## unreleased
+
+* Add `InterruptibilityState::set_fn_interrupt_poll_item` to run arbitrary function when interruption is activated and the underlying `Stream` returns `Poll::Ready(..)`.
+* Add tests to prove correctness of `PollNextN` `InterruptStrategy`.
+
+
 ## 0.1.0 (2024-01-02)
 
 * Add `InterruptibilityState::set_fn_interrupt_activate` to run arbitrary function when interruption is activated.